+++ Sources/BraintreeCore/Analytics/BTAnalyticsService.swift
@@ -31,6 +31,8 @@ class BTAnalyticsService: Equatable {
 
     private let apiClient: BTAPIClient
 
+    private var payPalContextID: String?
+
     // MARK: - Initializer
 
     init(apiClient: BTAPIClient, flushThreshold: Int = 1) {
@@ -45,8 +47,14 @@ class BTAnalyticsService: Equatable {
     ///  Events are queued and sent in batches to the analytics service, based on the status of the app and
     ///  the number of queued events. After exiting this method, there is no guarantee that the event has been sent.
     /// - Parameter eventName: String representing the event name
-    func sendAnalyticsEvent(_ eventName: String, errorDescription: String? = nil, correlationID: String? = nil) {
+    func sendAnalyticsEvent(
+        _ eventName: String,
+        errorDescription: String? = nil,
+        correlationID: String? = nil,
+        payPalContextID: String? = nil
+    ) {
         DispatchQueue.main.async {
+            self.payPalContextID = payPalContextID
             self.enqueueEvent(eventName, errorDescription: errorDescription, correlationID: correlationID)
             self.flushIfAtThreshold()
         }
@@ -57,9 +65,11 @@ class BTAnalyticsService: Equatable {
         _ eventName: String,
         errorDescription: String? = nil,
         correlationID: String? = nil,
+        payPalContextID: String? = nil,
         completion: @escaping (Error?) -> Void = { _ in }
     ) {
         DispatchQueue.main.async {
+            self.payPalContextID = payPalContextID
             self.enqueueEvent(eventName, errorDescription: errorDescription, correlationID: correlationID)
             self.flush(completion)
         }
@@ -156,6 +166,7 @@ class BTAnalyticsService: Equatable {
             environment: config.fptiEnvironment,
             integrationType: apiClient.metadata.integration.stringValue,
             merchantID: config.merchantID,
+            payPalContextID: payPalContextID,
             sessionID: sessionID,
             tokenizationKey: apiClient.tokenizationKey
         )