+++ Sources/BraintreeCore/BTAPIClient.swift
@@ -373,11 +373,17 @@ import Foundation
 
     /// :nodoc: This method is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     @_documentation(visibility: private)
-    public func sendAnalyticsEvent(_ eventName: String, errorDescription: String? = nil, correlationID: String? = nil) {
+    public func sendAnalyticsEvent(
+        _ eventName: String,
+        errorDescription: String? = nil,
+        correlationID: String? = nil,
+        payPalContextID: String? = nil
+    ) {
         analyticsService?.sendAnalyticsEvent(
             eventName,
             errorDescription: errorDescription,
             correlationID: correlationID,
+            payPalContextID: payPalContextID,
             completion: { _ in }
         )
     }